codeclimate-sync/requirements.txt
asn1crypto==1.0.0
astroid==2.3.1
boto3==1.9.242
botocore==1.12.242
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Deprecated==1.2.6
docutils==0.15.2
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
PyGithub==1.43.8
PyJWT==1.7.1
pylint==2.4.2
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
